+++ run-idea-2019.1.bash
@@ -1,2 +1,0 @@
-#!/usr/bin/env bash
-bash ./gradlew -PideaVersion="2019.1.3" -PideaBranch="191" runIde