+++ run-idea-2019.3.bash
@@ -1,0 +1,2 @@
+#!/usr/bin/env bash
+bash ./gradlew -PideaBranch=193 -PideaVersion=193.5233.12-EAP-SNAPSHOT runIde