+++ src/com/ansorgit/plugins/bash/lang/psi/api/vars/BashVarDef.java
@@ -23,6 +23,7 @@ import com.intellij.psi.PsiElement;
 import com.intellij.psi.PsiNameIdentifierOwner;
 import com.intellij.psi.PsiNamedElement;
 import org.jetbrains.annotations.NotNull;
+import org.jetbrains.annotations.Nullable;
 
 /**
  * @author jansorg
@@ -82,7 +83,7 @@ public interface BashVarDef extends BashPsiElement, PsiNamedElement, PsiNameIden
      */
     boolean hasAssignmentValue();
 
-    @NotNull
+    @Nullable
     BashReference getReference();
 
     /**