+++ src/com/ansorgit/plugins/bash/lang/psi/impl/vars/DumbVarDefReference.java
@@ -22,7 +22,7 @@ import org.jetbrains.annotations.Nullable;
 /**
  * @author jansorg
  */
-class DumbVarDefReference extends AbstractVarDefReference {
+public class DumbVarDefReference extends AbstractVarDefReference {
     public DumbVarDefReference(BashVarDefImpl bashVarDef) {
         super(bashVarDef);
     }